Added Edit CashBill

+++ DienMayQuyetTien/Areas/Admin/Controllers/CashBillsController.cs
@@ -20,6 +20,7 @@ namespace DienMayQuyetTien.Areas.Admin.Controllers
         {
             if (Session["UserName"] != null)
             {
+                ViewBag.Message = TempData["message"];
                 return View(db.CashBills.ToList());
             }
             else
@@ -105,6 +106,8 @@ namespace DienMayQuyetTien.Areas.Admin.Controllers
 
                     Session["CashBill"] = null;
                     Session["CashBillDetail"] = null;
+                    Session["total"] = null;
+                    TempData["message"] = "Tạo hóa đơn thành công.";
                     return RedirectToAction("Index");
                 }
                 catch (Exception e)
@@ -126,6 +129,7 @@ namespace DienMayQuyetTien.Areas.Admin.Controllers
             {
                 return HttpNotFound();
             }
+            Session["CashBill"] = cashBill;
             return View(cashBill);
         }
 
@@ -134,15 +138,52 @@ namespace DienMayQuyetTien.Areas.Admin.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,BillCode,CustomerName,PhoneNumber,Address,Date,Shipper,Note,GrandTotal")] CashBill cashBill)
+        public ActionResult Edit(CashBill cashBill)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(cashBill).State = EntityState.Modified;
+                Session["CashBill"] = cashBill;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return Redirect(ControllerContext.HttpContext.Request.UrlReferrer.ToString());
             }
             return View(cashBill);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Edit2()
+        {
+            using (var scope = new TransactionScope())
+                try
+                {
+                    var cashBill = Session["CashBill"] as CashBill;
+                    var CTHoaDon = Session["CashBillDetail"] as List<CashBillDetail>;
+                    cashBill.Date = DateTime.Now;
+                    cashBill.GrandTotal = (int)Session["total"];
+
+                    db.Entry(cashBill).State = EntityState.Modified;
+                    db.SaveChanges();
+
+                    foreach (var chiTiet in CTHoaDon)
+                    {
+                        chiTiet.BillID = cashBill.ID;
+                        chiTiet.Product = null;
+                        db.CashBillDetails.Add(chiTiet);
+                    }
+                    db.SaveChanges();
+                    scope.Complete();
+                    
+                    Session["CashBillDetail"] = null;
+                    Session["total"] = null;
+                    TempData["message"] = "Chỉnh sửa hóa đơn thành công.";
+                    return RedirectToAction("Index");
+                }
+                catch (Exception e)
+                {
+                    ModelState.AddModelError("", e.Message);
+                }
+            return View("Edit");
         }
 
         // GET: Admin/CashBills/Delete/5
@@ -165,10 +206,43 @@ namespace DienMayQuyetTien.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            CashBill cashBill = db.CashBills.Find(id);
-            db.CashBills.Remove(cashBill);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (db.CashBillDetails.Where(c => c.BillID == id).ToList().Count > 0 || db.InstallmentBillDetails.Where(i => i.BillID == id).ToList().Count > 0)
+            {
+                TempData["message"] = "Không thể xóa hóa đơn có chứa sản phẩm.";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                CashBill cashBill = db.CashBills.Find(id);
+                db.CashBills.Remove(cashBill);
+                db.SaveChanges();
+                TempData["message"] = "Xóa hóa đơn thành công.";
+                return RedirectToAction("Index");
+            }
+
+        }
+
+        public PartialViewResult Print(int id)
+        {
+            var cashbill = db.CashBills.FirstOrDefault(c => c.ID == id);
+            if (cashbill != null)
+            {
+                ReceiptCBModel rp = new ReceiptCBModel();
+                rp.Address = cashbill.Address;
+                rp.BillCode = cashbill.BillCode;
+                rp.CustomerName = cashbill.CustomerName;
+                rp.Date = cashbill.Date;
+                rp.cashBillDetail = cashbill.CashBillDetails.ToList();
+                rp.Shipper = cashbill.Shipper;
+                rp.PhoneNumber = cashbill.PhoneNumber;
+                rp.GrandTotal = cashbill.GrandTotal;
+                rp.Note = cashbill.Note;
+                return PartialView(rp);
+            }
+            else
+            {
+                return PartialView();
+            }
         }
 
         protected override void Dispose(bool disposing)
